Backend/requirements.txt
pytest==7.0.1
astroid==2.9.3  
attrs==21.4.0 
click==8.0.4  
Flask==2.0.3
Flask_Cors==3.0.10
iniconfig==1.1.1  
isort==5.10.1 
itsdangerous==2.1.1  
Jinja2==3.0.3  
lazy-object-proxy==1.7.1  
MarkupSafe==2.1.0  
mccabe==0.6.1  
packaging==21.3   
pip==21.1 
platformdirs==2.5.1  
pluggy==1.0.0  
py==1.11.0 
pylint==2.12.2 
pyparsing==3.0.7  
pytest==7.0.1  
setuptools==44.0.0 
toml==0.10.2 
tomli==2.0.1  
typing-extensions==4.1.1  
Werkzeug==2.0.3
wrapt==1.13.3
PyJWT==2.3.0
fpdf==1.7.2
# pdfplumber==0.6.0
json2html~=1.3.0
requests == 2.25.1
gunicorn==20.0.4